Fix ASIO HW combobox dropdown size
In MFC, the combobox height define the height of the combobox + its
dropdown list, as said here [0]:

> The height parameter applies to the height of a combo box with a
> drop-down list fully expanded.

The height wasn't enough to old one item and could cause the dropdown list
to be empty despite still having items.

So increase it so at least some items are shown.

[0] https://docs.microsoft.com/en-us/windows/desktop/menurc/combobox-control

Fix: eiz#56